adreno_tz: Fix GPU target frequency calculation for high refresh rates
It seems that the black-box GPU frequency selection algorithm is unaware
of high refresh rates, and assumes a refresh rate of 60 Hz for its
target frequencies. This results in frequencies that are too low being
selected when a higher refresh rate, such as 90 Hz, is used.

To remedy this, multiply the GPU's busy time by a ratio of the current
refresh rate to 60 Hz. This makes msm-adreno-tz select the right
frequencies needed to avoid stuttering.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>

[RyuujiX]: Due me using custom refresh rate with cmdline [1], so take values
from refresh_rate_cus.

[1] https://github.com/RyuujiX/android_kernel_asus_sdm660-4.19/commit/5eaf67969c5d97eaad2ae79865476abbf0919c50

Co-authored-by: RyuujiX <saputradenny712@gmail.com>
Signed-off-by: RyuujiX <saputradenny712@gmail.com>
Signed-off-by: dotkit <dotkit@electrowizard.me>