msm: kgsl: Don't allocate memory dynamically for temp command buffers
The temporary command buffer in _set_pagetable_gpu is only the size of a
single page, and _set_pagetable_gpu is never executed concurrently. It
is therefore easy to replace the dynamic command buffer allocation with
a static one to improve performance by avoiding the latency of dynamic
memory allocation.

Change-Id: I20bdf47c26e7f0b8ff4f569ac95c2be4ad69a9a3
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Kneba <abenkenary3@gmail.com>